refactor: update import paths for ThemedText to follow naming conventions
